pyproject.toml
[project]
name = "lirik-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.29.5",
    "polars>=1.36.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.9",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
